Move the Ouroboros mini-protocols mock server in its own (test) module
